pyproject.toml
[tool.poetry]
name = "chemistry"
version = "0.1.0"
description = ""
authors = ["mn-thermodynamics <mn@heaten.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
openpyxl = "^3.1.2"
fastapi = "^0.110.1"
pydantic = "^2.7.0"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
